fix fs-cleanup

+++ fs-cleanup.sh
@@ -19,5 +19,8 @@ rm "$1/lib32"
 echo "nuking unnecessary files"
 rm "$1/usr/lib/os-release"
 
-echo "Fixing broken magic.mgc"
-cp magic.mgc "$1/usr/share/misc/magic.mgc"
+if [ -f "$1/usr/share/magic/magic.mgc" ]; then
+    echo "Fixing broken magic.mgc"
+    cp magic.mgc "$1/usr/share/misc/magic.mgc"
+fi
+echo "done!"
